moved user commands under organization
