+++ cloud/user/user_test.go
@@ -304,7 +304,7 @@ func TestListOrgUser(t *testing.T) {
 		out := new(bytes.Buffer)
 		mockClient := new(astrocore_mocks.ClientWithResponsesInterface)
 		mockClient.On("ListOrgUsersWithResponse", mock.Anything, mock.Anything, mock.Anything).Return(&ListOrgUsersResponseOK, nil).Twice()
-		err := ListOrgUsers(out, mockClient, 1)
+		err := ListOrgUsers(out, mockClient)
 		assert.NoError(t, err)
 	})
 
@@ -312,7 +312,7 @@ func TestListOrgUser(t *testing.T) {
 		out := new(bytes.Buffer)
 		mockClient := new(astrocore_mocks.ClientWithResponsesInterface)
 		mockClient.On("ListOrgUsersWithResponse", mock.Anything, mock.Anything, mock.Anything).Return(nil, errorNetwork).Once()
-		err := ListOrgUsers(out, mockClient, 1)
+		err := ListOrgUsers(out, mockClient)
 		assert.EqualError(t, err, "network error")
 	})
 
@@ -320,7 +320,7 @@ func TestListOrgUser(t *testing.T) {
 		out := new(bytes.Buffer)
 		mockClient := new(astrocore_mocks.ClientWithResponsesInterface)
 		mockClient.On("ListOrgUsersWithResponse", mock.Anything, mock.Anything, mock.Anything).Return(&ListOrgUsersResponseError, nil).Twice()
-		err := ListOrgUsers(out, mockClient, 1)
+		err := ListOrgUsers(out, mockClient)
 		assert.EqualError(t, err, "failed to list users")
 	})
 
@@ -332,7 +332,7 @@ func TestListOrgUser(t *testing.T) {
 		assert.NoError(t, err)
 		out := new(bytes.Buffer)
 		mockClient := new(astrocore_mocks.ClientWithResponsesInterface)
-		err = ListOrgUsers(out, mockClient, 1)
+		err = ListOrgUsers(out, mockClient)
 		assert.ErrorIs(t, err, ErrNoShortName)
 	})
 
@@ -341,7 +341,7 @@ func TestListOrgUser(t *testing.T) {
 		expectedOutMessage := ""
 		out := new(bytes.Buffer)
 		mockClient := new(astrocore_mocks.ClientWithResponsesInterface)
-		err := ListOrgUsers(out, mockClient, 1)
+		err := ListOrgUsers(out, mockClient)
 		assert.Error(t, err)
 		assert.Equal(t, expectedOutMessage, out.String())
 	})
